Fix error with unsigned

+++ reader.go
@@ -229,7 +229,7 @@ func (r *Reader) ReadQBitArray() ([]bool, error) {
 		return nil, err
 	}
 	bits := make([]bool, n)
-	for i := n - 1; i >= 0; i-- {
+	for i := int64(n - 1); i >= 0; i-- {
 		bits[i] = (((buf[i/8]) >> (7 - i%8)) & 0x1) == 0x1
 	}
 	return bits, nil
